third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/blur-while-edit-as-html-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that HTML editor hides only when focusing another element


Running: testSetUp

Running: selectNode

Running: testBlurWithoutRelatedTarget
Active element: TEXTAREA
Is editing: true

Running: testBlurWithRelatedTarget
Active element: TEXTAREA
Is editing: false
